JSON parse

+++ db/migrate/20230203131730_create_cards.rb
@@ -1,7 +1,7 @@
 class CreateCards < ActiveRecord::Migration[7.0]
   def change
     create_table :cards do |t|
-      t.float :bonuses, default: 0
+      t.integer :bonuses, default: 0
       t.references :user, foreign_key: true
       t.references :shop, foreign_key: true
